requirements.txt
Flask
gunicorn
requests
python-dotenv
replicate